refactor: rename packages check_box to checkBox
